Change the AfterFx IPC model to a blocking listen
This is required because on Windows Server the async calls such as
`setTimeout` and `scheduleTask` never seem to trigger.

In this implementation we have a single blocking task that listens for
messages and responds to them.
This means that we have to only control AfterEffects via our IPC channel
and cannot spawn `aerender.exe` instances to control AfterEffects (they
will hang).

We have also updated the IPC host address to `localhost` instead.

Signed-off-by: Alex Hughes <ahughesalex@gmail.com>